+++ app/src/main/java/com/example/pinkorange/vibrato/LiveWithSettings.java
@@ -1,7 +1,9 @@
 package com.example.pinkorange.vibrato;
 
+import android.Manifest;
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.media.AudioAttributes;
 import android.media.AudioFormat;
 import android.media.AudioRecord;
@@ -11,6 +13,8 @@ import android.media.MediaRecorder;
 import android.media.audiofx.BassBoost;
 import android.media.audiofx.LoudnessEnhancer;
 import android.os.Bundle;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
 import android.util.Log;
 import android.support.design.widget.NavigationView;
 import android.support.v4.view.GravityCompat;
@@ -72,6 +76,7 @@ public class LiveWithSettings extends AppCompatActivity
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate (savedInstanceState);
         setContentView(R.layout.activity_music);
+        requestPermissions();
         initVariables();
 
         setSupportActionBar(toolbar);
@@ -381,5 +386,24 @@ public class LiveWithSettings extends AppCompatActivity
                 }
             } while (isRecording);
         }
+    }
+
+     private void requestPermissions() {
+        if (ContextCompat.checkSelfPermission(getApplicationContext(), Manifest.permission.MODIFY_AUDIO_SETTINGS) == PackageManager.PERMISSION_DENIED)
+            Log.d("App", "No MODIFY_AUDIO_SETTINGS" );
+        else
+            Log.d("App", "Yes MODIFY_AUDIO_SETTINGS" );
+        if (ContextCompat.checkSelfPermission(getApplicationContext(), Manifest.permission.RECORD_AUDIO) == PackageManager.PERMISSION_DENIED)
+            Log.d("App", "No RECORD_AUDIO" );
+        else
+            Log.d("App", "Yes RECORD_AUDIO" );
+
+        Log.d("App","Requesting permissions" );
+        ActivityCompat.requestPermissions( this, new String[]
+                {
+                        Manifest.permission.MODIFY_AUDIO_SETTINGS,
+                        Manifest.permission.RECORD_AUDIO
+                },1 );
+        Log.d("App","Requested perms");
     }
 }
