+++ app/src/main/java/com/example/pinkorange/vibrato/MainActivity.java
@@ -1,10 +1,6 @@
 package com.example.pinkorange.vibrato;
 
-import android.Manifest;
 import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
@@ -18,8 +14,6 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        requestPermissions();
 
         Button recorded_music_button = findViewById(R.id.recorded_music_button);
         recordedIntent = new Intent(MainActivity.this, SelectMusic.class);
@@ -40,23 +34,6 @@ public class MainActivity extends AppCompatActivity {
          });
     }
 
-    private void requestPermissions() {
-        if (ContextCompat.checkSelfPermission(getApplicationContext(), Manifest.permission.MODIFY_AUDIO_SETTINGS) == PackageManager.PERMISSION_DENIED)
-            Log.d("App", "No MODIFY_AUDIO_SETTINGS" );
-        else
-            Log.d("App", "Yes MODIFY_AUDIO_SETTINGS" );
-        if (ContextCompat.checkSelfPermission(getApplicationContext(), Manifest.permission.RECORD_AUDIO) == PackageManager.PERMISSION_DENIED)
-            Log.d("App", "No RECORD_AUDIO" );
-        else
-            Log.d("App", "Yes RECORD_AUDIO" );
 
-        Log.d("App","Requesting permissions" );
-        ActivityCompat.requestPermissions( this, new String[]
-                {
-                        Manifest.permission.MODIFY_AUDIO_SETTINGS,
-                        Manifest.permission.RECORD_AUDIO
-                },1 );
-        Log.d("App","Requested perms");
-    }
 
 }
